- "Import Profile" is now asking to overwrite an existing profile
When importing a profile there is now a check in place to ensure the user is aware that he/she is overwriting an existing profile and given the chance to change the profile name.
In addition all profiles (existing and new) are reset to default values prior to the import. This should fix the broken mods issue (unecessary events with code from other mods) which occurred during import for several users and will ensure newly added default values, which are not present in the imported profile, are populated correctly.